Internal: Address review feedback on component instance MCP tools
Merge list-component-schemas into list-components as a two-step tool
mirroring list-widget-schemas: calling without arguments lists components
without their overridable props, so sites with many components are not
forced to build every schema up front.

Share the component-instance envelope reader via the prop type that owns
the shape, replacing four duplicated lookups, and build envelopes from the
prop type keys instead of hardcoded $$type strings.

Drop the component instance shape from the manage-elements input schema
description; it already lives in the list-components prompt.
