+++ modules/components/widgets/component-instance.php
@@ -123,12 +123,6 @@ class Component_Instance extends Atomic_Widget_Base {
 	}
 
 	private function get_component_id(): ?int {
-		$settings = $this->get_settings();
-
-		if ( ! isset( $settings['component_instance']['value']['component_id']['value'] ) ) {
-			return null;
-		}
-
-		return (int) $settings['component_instance']['value']['component_id']['value'];
+		return Component_Instance_Prop_Type::extract_component_id( $this->get_settings() );
 	}
 }
