+++ tests/phpunit/elementor/modules/mcp/test-list-components-ability.php
@@ -6,11 +6,14 @@ use Elementor\Modules\Components\Components_Repository;
 use Elementor\Modules\Components\Documents\Component as Component_Document;
 use Elementor\Modules\Mcp\Abilities\List_Components_Ability;
 use Elementor\Plugin;
+use Elementor\Testing\Modules\Components\Mocks\Component_Overrides_Mocks;
 use ElementorEditorTesting\Elementor_Test_Base;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
+
+require_once __DIR__ . '/../components/mocks/component-overrides-mocks.php';
 
 /**
  * @group Elementor\Modules\Mcp
@@ -70,6 +73,20 @@ class Test_List_Components_Ability extends Elementor_Test_Base {
 		$this->assertFalse( $component['is_archived'] );
 	}
 
+	public function test_execute__omits_overridable_props_from_the_discovery_list() {
+		// Arrange
+		$this->act_as_admin();
+		$this->create_component_with_overridable_props();
+		$ability = new List_Components_Ability();
+
+		// Act
+		$result = $ability->execute();
+
+		// Assert
+		$this->assertCount( 1, $result['components'] );
+		$this->assertArrayNotHasKey( 'overridable_props', $result['components'][0] );
+	}
+
 	public function test_execute__returns_archived_flag_for_archived_components() {
 		// Arrange
 		$this->act_as_admin();
@@ -87,9 +104,155 @@ class Test_List_Components_Ability extends Elementor_Test_Base {
 		$this->assertTrue( $result['components'][0]['is_archived'] );
 	}
 
+	public function test_execute__returns_discovery_list_when_component_ids_is_empty_array() {
+		// Arrange
+		$this->act_as_admin();
+		$component_id = $this->create_component( 'My Hero', 'publish' );
+		$ability = new List_Components_Ability();
+
+		// Act
+		$result = $ability->execute( [ 'component_ids' => [] ] );
+
+		// Assert
+		$this->assertCount( 1, $result['components'] );
+		$this->assertSame( $component_id, $result['components'][0]['id'] );
+		$this->assertArrayNotHasKey( 'overridable_props', $result['components'][0] );
+	}
+
+	public function test_execute__returns_400_when_component_ids_contains_only_invalid_values() {
+		// Arrange
+		$this->act_as_admin();
+		$ability = new List_Components_Ability();
+
+		// Act
+		$result = $ability->execute( [ 'component_ids' => [ 0, -1, 'abc' ] ] );
+
+		// Assert
+		$this->assertWPError( $result );
+		$this->assertSame( 'invalid_input', $result->get_error_code() );
+		$this->assertSame( \WP_Http::BAD_REQUEST, $result->get_error_data()['status'] );
+	}
+
+	public function test_execute__returns_404_when_any_requested_component_is_missing() {
+		// Arrange
+		$this->act_as_admin();
+		$existing_id = $this->create_component( 'Existing', 'publish' );
+		$ability = new List_Components_Ability();
+
+		// Act
+		$result = $ability->execute( [ 'component_ids' => [ $existing_id, 999999 ] ] );
+
+		// Assert
+		$this->assertWPError( $result );
+		$this->assertSame( 'elementor_not_found', $result->get_error_code() );
+		$this->assertSame( \WP_Http::NOT_FOUND, $result->get_error_data()['status'] );
+	}
+
+	public function test_execute__returns_404_when_id_points_to_non_component_post() {
+		// Arrange
+		$this->act_as_admin();
+		$post_id = $this->factory()->post->create();
+		$ability = new List_Components_Ability();
+
+		// Act
+		$result = $ability->execute( [ 'component_ids' => [ $post_id ] ] );
+
+		// Assert
+		$this->assertWPError( $result );
+		$this->assertSame( 'elementor_not_found', $result->get_error_code() );
+	}
+
+	public function test_execute__returns_schema_with_empty_overridable_props() {
+		// Arrange
+		$this->act_as_admin();
+		$component_id = $this->create_component( 'Card Component', 'publish' );
+		$ability = new List_Components_Ability();
+
+		// Act
+		$result = $ability->execute( [ 'component_ids' => [ $component_id ] ] );
+
+		// Assert
+		$this->assertCount( 1, $result['components'] );
+
+		$entry = $result['components'][0];
+		$this->assertSame( $component_id, $entry['id'] );
+		$this->assertSame( 'Card Component', $entry['name'] );
+		$this->assertArrayHasKey( 'uid', $entry );
+		$this->assertFalse( $entry['is_archived'] );
+		$this->assertIsArray( $entry['overridable_props'] );
+		$this->assertEmpty( $entry['overridable_props'] );
+	}
+
+	public function test_execute__returns_multiple_components_in_input_order() {
+		// Arrange
+		$this->act_as_admin();
+		$first_id = $this->create_component( 'First', 'publish' );
+		$second_id = $this->create_component( 'Second', 'publish' );
+		$ability = new List_Components_Ability();
+
+		// Act
+		$result = $ability->execute( [ 'component_ids' => [ $second_id, $first_id ] ] );
+
+		// Assert
+		$this->assertCount( 2, $result['components'] );
+		$this->assertSame( $second_id, $result['components'][0]['id'] );
+		$this->assertSame( $first_id, $result['components'][1]['id'] );
+	}
+
+	public function test_execute__deduplicates_repeated_component_ids() {
+		// Arrange
+		$this->act_as_admin();
+		$component_id = $this->create_component( 'Solo', 'publish' );
+		$ability = new List_Components_Ability();
+
+		// Act
+		$result = $ability->execute( [ 'component_ids' => [ $component_id, $component_id ] ] );
+
+		// Assert
+		$this->assertCount( 1, $result['components'] );
+		$this->assertSame( $component_id, $result['components'][0]['id'] );
+	}
+
+	public function test_execute__returns_overridable_props_schema_with_labels_and_groups() {
+		// Arrange
+		$this->act_as_admin();
+		$component_id = $this->create_component_with_overridable_props();
+		$ability = new List_Components_Ability();
+
+		// Act
+		$result = $ability->execute( [ 'component_ids' => [ $component_id ] ] );
+
+		// Assert
+		$this->assertIsArray( $result['components'][0]['overridable_props'] );
+		$this->assertArrayHasKey( 'prop-uuid-1', $result['components'][0]['overridable_props'] );
+
+		$title_prop = $result['components'][0]['overridable_props']['prop-uuid-1'];
+		$this->assertSame( 'Heading Title', $title_prop['label'] );
+		$this->assertSame( 'group-1', $title_prop['group_id'] );
+		$this->assertArrayHasKey( 'origin_prop_schema', $title_prop );
+
+		$origin_schema = $title_prop['origin_prop_schema'];
+		$this->assertArrayNotHasKey( '$$type', $origin_schema['properties'] ?? [],
+			'origin_prop_schema must be plain-value JSON schema (no $$type envelope), matching elementor/get-widget-schema.' );
+		$this->assertStringNotContainsString( '"$$type"', wp_json_encode( $origin_schema ),
+			'origin_prop_schema must not contain any $$type envelopes anywhere in the tree.' );
+	}
+
 	private function create_component( string $title, string $status ): int {
 		$repository = new Components_Repository();
 		return $repository->create( $title, [], $status, uniqid( 'uid-', true ) );
+	}
+
+	private function create_component_with_overridable_props(): int {
+		$component_id = $this->create_component( 'Hero Component', 'publish' );
+
+		$repository = new Components_Repository();
+		$component = $repository->get( $component_id, false );
+
+		$mocks = new Component_Overrides_Mocks();
+		$component->update_overridable_props( $mocks->get_mock_component_overridable_props() );
+
+		return $component_id;
 	}
 
 	private function delete_all_components(): void {
